Reorganized Drivetrain and TankDrive
