+++ Code2017/src/org/usfirst/frc253/Code2017/subsystems/Drivetrain.java
@@ -1,8 +1,8 @@
 
-package org.usfirst.frc253.Code2017.commands;
+package org.usfirst.frc253.Code2017.subsystems;
 
 import org.usfirst.frc253.Code2017.RobotMap;
-import org.usfirst.frc253.Code2017.subsystems.TankDrive;
+import org.usfirst.frc253.Code2017.commands.TankDrive;
 
 import edu.wpi.first.wpilibj.Talon;
 import edu.wpi.first.wpilibj.Victor;
